MainPage 컴포넌트 내부에 있던 ImageCardItem 컴포넌트를 외부 파일로 분리
