+++ src/pages/index.tsx
@@ -1,15 +1,12 @@
-import { useRouter } from 'next/router'
 import { useSWRConfig } from 'swr'
 import withHeader from 'hocs/withHeader'
 import withLogin from 'hocs/withLogin'
-import { css } from '@emotion/react'
 import ImageList from '@mui/material/ImageList'
-import ImageListItem from '@mui/material/ImageListItem'
-import ImageListItemBar from '@mui/material/ImageListItemBar'
 import LikeButton from 'components/common/buttons/likeButton'
 import useMe from 'hooks/useMe'
 import useFilteredSnsPosts from 'hooks/useFilteredSnsPosts'
 import createUrlQuery from 'utils/createUrlQuery'
+import ImageCardItem from 'components/home/imageCardItem'
 
 const query = createUrlQuery({
   'populate[0]': 'postImages',
@@ -18,42 +15,7 @@ const query = createUrlQuery({
   'pagination[limit]': 200,
 })
 
-const cursorPointer = css`
-  cursor: pointer;
-`
-
 const SnsPostLikeButtonWithLogin = withLogin(LikeButton)
-
-const ImageCardItem = ({ cardItemData, rightActionButton }) => {
-  const router = useRouter()
-
-  const goToSnsPost = (snsPostId: number) =>
-    router.push(`/sns/post/${snsPostId}`)
-
-  const handleImageListItemClick = (e: any) => {
-    const isLikeButtonClicked = e.target.id === LikeButton.id
-    if (isLikeButtonClicked) {
-      return
-    }
-
-    goToSnsPost(cardItemData.id)
-  }
-
-  return (
-    <ImageListItem onClick={handleImageListItemClick} css={cursorPointer}>
-      <img
-        src={cardItemData.postImage.url}
-        alt={cardItemData.postImage.altText}
-      />
-      <ImageListItemBar
-        title={cardItemData.author.username}
-        position="top"
-        actionIcon={rightActionButton}
-        actionPosition="right"
-      />
-    </ImageListItem>
-  )
-}
 
 const MainPage = () => {
   const { mutate } = useSWRConfig()
